requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.2